+++ app/js/homeydash.helpers.js
@@ -42,7 +42,7 @@ var setLocale = function () {
 
 function setCookie(cname, cvalue, exweeks) {
   var d = new Date();
-  d.setTime(d.getTime() + (exweeks*7*24*60*60*1000));
+  d.setTime(d.getTime() + (exweeks*5*7*24*60*60*1000));
   var expires = "expires="+ d.toUTCString();
   document.cookie = cname + "=" + cvalue + ";" + expires + ";path=/";
 }
@@ -77,4 +77,11 @@ function dynamicSort(property) {
 
 function setScale(scale) {
   window.body.style.zoom = scale/10;
+}
+
+function checkImage (src, good, bad) {
+  var img = new Image();
+  img.onload = good; 
+  img.onerror = bad;
+  img.src = src;
 }